LayoutTests/fast/dom/HTMLAnchorElement/get-parameter-expected.txt

Test calling getParameter without required arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a.getParameter() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
